fix code and linting format
