+++ lib/features/file/browser/utils/empty_directory_view.dart
@@ -51,7 +51,7 @@ class EmptyDirectoryView extends StatelessWidget {
           Icon(
             Icons.folder_open,
             size: 48,
-            color: Theme.of(context).colorScheme.primary.withOpacity(0.6),
+            color: Theme.of(context).colorScheme.primary.withValues(alpha: 0.6),
           ),
 
           const SizedBox(height: 16),
@@ -65,7 +65,9 @@ class EmptyDirectoryView extends StatelessWidget {
               ).colorScheme.primary.withValues(alpha: 0.1),
               borderRadius: BorderRadius.circular(AppTheme.defaultBorderRadius),
               border: Border.all(
-                color: Theme.of(context).colorScheme.outline.withOpacity(0.2),
+                color: Theme.of(
+                  context,
+                ).colorScheme.outline.withValues(alpha: 0.2),
               ),
             ),
             child: Text(
