+++ lib/screens/settings_screen.dart
@@ -26,8 +26,8 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:moviestar/screens/to_watch_screen.dart';
