+++ lib/utils/create_solid_login.dart
@@ -67,42 +67,39 @@ Widget _buildNormalLogin(String serverUrl, SharedPreferences prefs) {
     builder: (context) {
       // Wrap SolidLogin in a container with custom image.
 
-      return Container(
-        // Let SolidLogin handle its own theming
-        child: Column(
-          children: [
-            Expanded(
-              child: Theme(
-                data:
-                    Theme.of(context).brightness == Brightness.dark
-                        ? ThemeData.dark()
-                        : ThemeData.light(),
-                child: SolidLogin(
-                  required: false,
-                  title: 'Movie Star',
-                  appDirectory: 'moviestar',
-                  webID:
-                      serverUrl.isNotEmpty
-                          ? serverUrl
-                          : 'https://pods.dev.solidcommunity.au',
-                  image: const AssetImage('assets/images/app_image.jpg'),
-                  logo: const AssetImage('assets/images/app_icon.png'),
-                  link:
-                      'https://github.com/yourusername/moviestar/blob/main/README.md',
+      return Column(
+        children: [
+          Expanded(
+            child: Theme(
+              data:
+                  Theme.of(context).brightness == Brightness.dark
+                      ? ThemeData.dark()
+                      : ThemeData.light(),
+              child: SolidLogin(
+                required: false,
+                title: 'Movie Star',
+                appDirectory: 'moviestar',
+                webID:
+                    serverUrl.isNotEmpty
+                        ? serverUrl
+                        : 'https://pods.dev.solidcommunity.au',
+                image: const AssetImage('assets/images/app_image.jpg'),
+                logo: const AssetImage('assets/images/app_icon.png'),
+                link:
+                    'https://github.com/yourusername/moviestar/blob/main/README.md',
 
-                  // Use a wrapper widget to check for API key after login
-                  child: ApiKeyCheckWrapper(
+                // Use a wrapper widget to check for API key after login.
+                child: ApiKeyCheckWrapper(
+                  prefs: prefs,
+                  child: MyHomePage(
+                    title: 'Movie Star Home Page',
                     prefs: prefs,
-                    child: MyHomePage(
-                      title: 'Movie Star Home Page',
-                      prefs: prefs,
-                    ),
                   ),
                 ),
               ),
             ),
-          ],
-        ),
+          ),
+        ],
       );
     },
   );
